v7r5p4: Setting the groundwork for very different new FEB firmware.
We do not read the FPGA's on every trigger, so the firmware version is not generally available
"dynamically."  Therefore we have to set it BY HAND.  Be sure that this is set correctly for
the *whole* detector.  Mixed firmware is pretty much disallowed.

Currently we are using v83 in the code (and on test stands).