Some comments from Cristian:
1. v90 will have 7+1 hits available.

2. The ‘preview hit’ described is implemented and works this way:
a. By default it is disabled.
b. To enable it, you need to set to ‘1’ the bit#7 (MSB) of the 32nd byte in the FPGA device register. To disable it, set this bit to ‘0’. The other bits of this register are used for reading the FEB address (bits 3 down to 0)  and for setting the number of HV averages (bits 6 down to 4). They are unchanged. For easy localization, the next byte in FPGA registers, the 33rd byte, is used to read the firmware version. See docdb 4311.
c. The working procedure I suggest is the following.
i) First Clear Status and reset DPM pointer.
ii) Then send an open gate fast command (it is assumed that the FEBs are ‘aligned’ such as the gate opens AND most important ends at the same time for all FEBs in a chain). At the end of the gate, each FEB will send a frame (which I’ll describe shortly).
iii) Then do again a Clear Status and reset DPM pointer. This is necessary because the CROC Channel has only one Status Register and what is happening now is that the CROC is no more the ‘master’ or initiator of frames: it receive a bunch of frames (the message received bit will be set by the ‘last’ FEB in the chain and then it will not ‘see’ the other responses) without sending any message (thus the message sent bit will not be set) and the frames received have a CRC error.
iv) Read the DPM. Since the pointer was reset (i) you start the reading from address zero. For example, if your chain contains 3 FEBs with addresses 1 (the first in chain), 5 and 10(the last in chain) the DPM content should be (in hex):
0600 8ANN HVHV 0600 85NN HVHV 0600 81NN HVHV   where:
0600: is the (standard) received message length, including itself. Always 6 bytes or 2 words (of 16 bits)
8ANN: the 8 represent the frame’s direction SlaveToMaster or FebToCroc if you like. Always an 8. The A is the hex address of the FEB. The two NNs are the hex numbers of hits in that FEB. The second N (LeastSignificantFourBits) represents the Trip0==Trip1 number of hits. The first N (MostSignificantFourBits) represents Trip2==Trip3 number of hits.
HVHV: this is the actual HV read value, LSByte first. For example a DPM reading of A861 translates to decimal 25000 ADC counts.

3. The ‘dump read’ function for the FPGA device registers is a new type of reading the registers. It works this way:
a. Use the same FPGA device register 0x20.
b. Use / implement a new function (dump read) with the code 0x03. Thus, the first byte after the start frame byte will be 0x23. Old values were 0x21 for write and 0x22 for read.
c. No data bytes are required. The message is a ‘short frame’ type, similar with the discriminator and the analog frames.
d. The response is identical with the existing read function (0x02), so the decoding should be unchanged.

4. To deal with the after pulse issue.
a. An ‘AfterPulseDelay’ was implemented. The values are 0 to 15 in 53MHz clock ticks.
b. The default is zero, no extra delay - the same as it is now.
c. To create a delay, use the LeastSignificantFourBits of the 28th register in FPGA device registers. This register was a read only register used for what I called ‘HVReadDummy’. Now it is a write - read register. The MostSignificantFourBits can be used for whatever data you want to write, if any. Currently they do not affect in any way the FPGA functionality.
d. The new discriminator acquisition cycle is the following:
i) 16 clocks for discriminator registering (unchanged).
ii) N clocks of wait time. No other discriminators are registered, even if they fire.
iii) 4 clocks for writing the time stamp and the quarter ticks (unchanged)
iv) 13 clocks for Trip Clock Toggling signals – misc Resets (unchanged). In fact this sub-cycle has 16 clocks, from which the first three overlaps the previous sub-cycle iii)
e. The discriminator response message is unchanged – the current parsing is unchanged.