v91.  C. Gingu comments: "This new v91 (made on my computer with Xilinx 10.1) is potentially more robust for encoded clock signal integrity. It is ‘looking’ at the  enc_clk in all 4 phases of the main FPGA’s clock and attempts to recover the encoded data. It also constrains the enc_clk for timing implementation."
"I believe I found a method to ‘catch’ at least some of the bad boards and we
should implement this when we release a board out for use. Basically it checks
(also) the FCMDERR flag in the FPGA register. I suggest you Gabe to check for
this one too (and also the corresponding SCMDERR – for frames, data signal)
since it seems to flag potential ‘missing gate’ problems even before the
’missing gate’ occurs."

(Bad boards here are boards that occasionally fail to decode the encoded clock
to get open gate commands.)