Expand use of logging facilities at the expense of std out. Will need to hide some of the logging messages behind precompiler statements for speed optimization someday (esp. for frame error checking).
Very significant change: updated the FrameTypes informaiton on the
frames header content to reflect the status word that comes with the
CROC-E frames. This impacts unpacking data with the old CROCs (basically,
they are incompatible). We should fix this (perhaps by adding in the
number of words supplied to frame header by the vme card channel
and not tracking that info in FrameTypes, but that will change
the enum in FrameTypes to represent offsets instead of indices).