Fix two bugs in the SetupHeaderData function for the channel, first, actually write to hardware, second, write in the correct byte order.
Put a check for sequencer readout completion into the readout worker.
It isn't clear if just sleeping might be a better strategy (i.e., will
it take longer to go read all the status registers? - pretty quickly
they'll all indicate the sequencer is done and then we are just wasting
time, we will have to check this).