Build minervadaq on SLF6. Installed Linux packages to build against - sqlite-devel, log4cpp-devel, and boost-devel. Added doxygen linux package. Added the D0TESTSTAND locale to setupdaqenv.sh.
The build system is configured to treat warnings as errors.
The most numerous issue was in the hardware header files which use a C language convention for typedef.
Here is the guilty code:
typedef enum xxx {
};

Here is the corrected code:
typedef enum xxx {
} xxx;

A few files did not include a system header file for a C function used in the source file.