Update the unpacking of the discriminator frame for v95. In v91 each of the four Trip chips used a 4-bit number for the number of hits.
In v95 Trips 0 and 1 have the same number of hits as do Trips 2 and 3.
The v91 space for hits (16-bits) is now divided in half; 8-bits for
Trip0/1 and 8 bits for Trip2/3.  This allows for over 15 hits per Trip.
In v95 there is a different space limit in the FEB of 23 hits.